refactor: rename user-home to UserHome
